# swagger-output/swagger-specification.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0",
    "description": "API for School Management System including authentication, academic year, subjects, exams, and mobile integrations",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "API Support",
      "url": "https://www.schoolmanagementsystem.com/support",
      "email": "support@schoolmanagementsystem.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Subject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Subject ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Subject name",
            "example": "Mathematics"
          },
          "shortName": {
            "type": "string",
            "description": "Short form of the subject name",
            "example": "Math"
          },
          "code": {
            "type": "string",
            "description": "Subject code",
            "example": "MATH101"
          },
          "description": {
            "type": "string",
            "description": "Subject description",
            "example": "Basic mathematical principles and concepts"
          },
          "classLevel": {
            "type": "string",
            "description": "Grade or class level",
            "example": "Grade 10"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the subject was created",
            "example": "2023-01-01T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the subject was last updated",
            "example": "2023-01-01T12:00:00.000Z"
          }
        }
      },
      "SubjectDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Subject ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Subject name",
            "example": "Mathematics"
          },
          "shortName": {
            "type": "string",
            "description": "Short form of the subject name",
            "example": "Math"
          },
          "code": {
            "type": "string",
            "description": "Subject code",
            "example": "MATH101"
          },
          "description": {
            "type": "string",
            "description": "Subject description",
            "example": "Basic mathematical principles and concepts"
          },
          "classLevel": {
            "type": "string",
            "description": "Grade or class level",
            "example": "Grade 10"
          },
          "teachers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "description": "Teacher ID"
                },
                "name": {
                  "type": "string",
                  "description": "Teacher name"
                },
                "email": {
                  "type": "string",
                  "description": "Teacher email"
                }
              }
            }
          },
          "subclasses": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "description": "Subclass ID"
                },
                "name": {
                  "type": "string",
                  "description": "Subclass name"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the subject was created",
            "example": "2023-01-01T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the subject was last updated",
            "example": "2023-01-01T12:00:00.000Z"
          }
        }
      },
      "CreateSubjectRequest": {
        "type": "object",
        "required": [
          "name",
          "shortName",
          "code",
          "classLevel"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Subject name",
            "example": "Physics"
          },
          "shortName": {
            "type": "string",
            "description": "Short form of the subject name",
            "example": "Phys"
          },
          "code": {
            "type": "string",
            "description": "Subject code",
            "example": "PHYS101"
          },
          "description": {
            "type": "string",
            "description": "Subject description",
            "example": "Introduction to basic principles of physics"
          },
          "classLevel": {
            "type": "string",
            "description": "Grade or class level",
            "example": "Grade 11"
          }
        }
      },
      "UpdateSubjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Subject name",
            "example": "Advanced Physics"
          },
          "shortName": {
            "type": "string",
            "description": "Short form of the subject name",
            "example": "AP"
          },
          "code": {
            "type": "string",
            "description": "Subject code",
            "example": "PHYS201"
          },
          "description": {
            "type": "string",
            "description": "Subject description",
            "example": "Advanced physics concepts including mechanics and thermodynamics"
          },
          "classLevel": {
            "type": "string",
            "description": "Grade or class level",
            "example": "Grade 12"
          }
        }
      },
      "AssignTeacherRequest": {
        "type": "object",
        "required": [
          "teacherId"
        ],
        "properties": {
          "teacherId": {
            "type": "integer",
            "description": "ID of the teacher to assign to the subject",
            "example": 5
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/subjects": {
      "get": {
        "summary": "Get all subjects",
        "description": "Retrieves a list of all subjects",
        "tags": [
          "Subjects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "classLevel",
            "schema": {
              "type": "string"
            },
            "description": "Filter subjects by class level"
          }
        ],
        "responses": {
          "200": {
            "description": "List of subjects retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subjects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Subject"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new subject",
        "description": "Creates a new subject with the provided details",
        "tags": [
          "Subjects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subject created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/subjects/{id}": {
      "get": {
        "summary": "Get subject details",
        "description": "Retrieves details of a specific subject by ID",
        "tags": [
          "Subjects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Subject ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Subject details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Subject not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update subject details",
        "description": "Updates an existing subject with the provided details",
        "tags": [
          "Subjects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Subject ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subject updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Subject not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a subject",
        "description": "Deletes a specific subject by ID",
        "tags": [
          "Subjects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Subject ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Subject deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Subject not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/subjects/{id}/teachers": {
      "post": {
        "summary": "Assign a teacher to a subject",
        "description": "Assigns a teacher to a specific subject",
        "tags": [
          "Subjects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Subject ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTeacherRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Teacher assigned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "subject": {
                      "$ref": "#/components/schemas/SubjectDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Subject or teacher not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}